refactor(core): move analyze_raw_range from edit.rs to analyze.rs
Implements issue #691. Moves the read-only analyze_raw_range function
from the mutation-focused edit module to analyze.rs where it belongs
with the rest of the analyze_* tool family.

Error strategy Option A: adds NotAFile, InvalidRange, and Io variants
to AnalyzeError so analyze.rs remains self-contained without importing
EditError. Updates match arms in the server crate accordingly.

Changes:
- Add NotAFile(PathBuf), InvalidRange{start,end,total}, and Io(#[from] std::io::Error)
  variants to AnalyzeError in analyze.rs
- Move analyze_raw_range function from edit.rs to analyze.rs with updated
  return type Result<AnalyzeRawOutput, AnalyzeError>
- Move all analyze_raw_range tests from edit.rs to analyze.rs
- Update pub use blocks in core/lib.rs: remove from edit, add to analyze
- Update match arms in crates/aptu-coder/src/lib.rs from EditError to AnalyzeError
- Fix edit.rs module doc to remove analyze_raw reference
- Remove unused AnalyzeRawOutput import from edit.rs
